chore: rename Processor->Worker
The thread worker rename to Executor.
