+++ projects/pgai/pgai/cli.py
@@ -183,7 +183,7 @@ async def async_run_vectorizer_worker(
     concurrency: int | None,
     exit_on_error: bool | None,
 ) -> None:
-    from .vectorizer import Processor
+    from .vectorizer import Worker
 
     # gracefully handle being asked to shut down
     signal.signal(signal.SIGINT, shutdown_handler)
@@ -193,7 +193,7 @@ async def async_run_vectorizer_worker(
         wrapper_class=structlog.make_filtering_bound_logger(get_log_level(log_level))
     )
 
-    processor = Processor(
+    processor = Worker(
         db_url,
         datetime.timedelta(seconds=poll_interval),
         once,
