+++ projects/pgai/pgai/vectorizer/vectorizer.py
@@ -166,7 +166,7 @@ class Vectorizer(BaseModel):
         concurrency = concurrency or self.config.processing.concurrency
         tasks = [
             asyncio.create_task(
-                Worker(
+                Executor(
                     db_url,
                     self,
                     features,
@@ -734,7 +734,7 @@ class UUIDEncoder(json.JSONEncoder):
         return json.JSONEncoder.default(self, o)
 
 
-class Worker:
+class Executor:
     """
     Responsible for processing items from the work queue and generating embeddings.
